Add DOTAbility files and refactor multiple classes
Added new files for DOTAbility, both .cpp and .h implementation. Updated categories in AmathProjectileAbility.h and changed the method casting of the InputComponent in AftermathPlayerController.cpp.
Additionally, the AftermathCharacterBase.cpp class was refactored to improve the death sequence and included the DOTAbility in the game logic. Removed the AbilityInputTagHeld method. A few empty lines in some other files were also cleaned up.