Implement QuestionWidget and update homing projectiles
Added QuestionWidget class with three editable text boxes and added it to the MainCharacter. Updated homing projectiles to only target enemies that are alive and within a 2000 unit range, and to switch off homing if no valid targets are found. Also generated a random math equation in MainCharacter as part of the new UI widget.